requirements.txt
fastapi
uvicorn[standard]
requests
PyYAML
pytest
beautifulsoup4
lxml